parandused: avaldused, õpetajate otsing

+++ hois_back/src/main/java/ee/hitsa/ois/web/ApplicationController.java
@@ -17,8 +17,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import ee.hitsa.ois.domain.application.Application;
+import ee.hitsa.ois.domain.school.School;
 import ee.hitsa.ois.domain.student.Student;
+import ee.hitsa.ois.enums.ApplicationStatus;
 import ee.hitsa.ois.enums.ApplicationType;
+import ee.hitsa.ois.exceptions.HoisBusinessRuleException;
 import ee.hitsa.ois.repository.SchoolRepository;
 import ee.hitsa.ois.service.ApplicationService;
 import ee.hitsa.ois.service.security.HoisUserDetails;
@@ -42,8 +45,9 @@ public class ApplicationController {
     private SchoolRepository schoolRepository;
 
     @GetMapping("/{id:\\d+}")
-    public ApplicationDto get(@WithEntity("id") Application application) {
-        return ApplicationDto.of(application);
+    public ApplicationDto get(HoisUserDetails user, @WithEntity("id") Application application) {
+        UserUtil.assertSameSchool(user, application.getStudent().getSchool());
+        return applicationService.get(user, application);
     }
 
     @GetMapping("")
@@ -53,20 +57,25 @@ public class ApplicationController {
 
     @PostMapping("")
     public ApplicationDto create(@Valid @RequestBody ApplicationForm applicationForm, HoisUserDetails user) {
-        return get(applicationService.create(user, applicationForm));
+        return get(user, applicationService.create(user, applicationForm));
     }
 
     @PutMapping("/{id:\\d+}")
-    public ApplicationDto update(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestBody = true) Application application,
-            @Valid @RequestBody ApplicationForm applicationForm) {
-        UserUtil.assertSameSchool(user, application.getStudent().getSchool());
-        return get(applicationService.save(application, applicationForm));
+    public ApplicationDto update(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestBody = true) Application application, @Valid @RequestBody ApplicationForm applicationForm) {
+        checkUpdateBusinessRules(user, application, applicationForm);
+        return get(user, applicationService.save(application, applicationForm));
     }
 
     @DeleteMapping("/{id:\\d+}")
     public void delete(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestBody = true) Application application) {
-        UserUtil.assertSameSchool(user, application.getStudent().getSchool());
-        applicationService.delete(application);
+        Student student = application.getStudent();
+        School school = application.getStudent().getSchool();
+        ApplicationStatus status = ApplicationStatus.valueOf(EntityUtil.getCode(application.getStatus()));
+        if((UserUtil.isSame(user, student) || UserUtil.isSchoolAdmin(user, school)) && ApplicationStatus.AVALDUS_STAATUS_KOOST.equals(status)) {
+            applicationService.delete(application);
+        } else {
+            throw new HoisBusinessRuleException(String.format("user %s is not allowed to delete application %d with status %s", user.getUsername(), application.getId(), status.name()));
+        }
     }
 
     @GetMapping("/student/{id:\\d+}/validAcademicLeave")
@@ -87,31 +96,73 @@ public class ApplicationController {
     @PutMapping("/{id:\\d+}/submit")
     public ApplicationDto submit(HoisUserDetails user, @WithEntity(value = "id") Application application) {
         Student student = application.getStudent();
-        if (UserUtil.isStudentRepresentative(user, student)) {
+        ApplicationStatus status = ApplicationStatus.valueOf(EntityUtil.getCode(application.getStatus()));
+
+        if (UserUtil.isStudentRepresentative(user, student) && ApplicationStatus.AVALDUS_STAATUS_KOOST.equals(status)) {
             applicationService.submit(user, application);
-        } else if (UserUtil.isSame(user, student) || UserUtil.isSchoolAdmin(user, student.getSchool())) {
+        } else if (UserUtil.isSame(user, student) || UserUtil.isSchoolAdmin(user, student.getSchool()) && ApplicationStatus.AVALDUS_STAATUS_KOOST.equals(status)) {
             applicationService.submit(user, application);
         } else {
-            throw new RuntimeException("User has no rights to submit this application");
+            throw new HoisBusinessRuleException(String.format("User %s is not allowed to submit application %d with status %s", user.getUsername(), application.getId(), status));
         }
 
-        return get(application);
+        return get(user, application);
     }
 
     @PutMapping("/{id:\\d+}/reject")
     public ApplicationDto reject(HoisUserDetails user, @WithEntity(value = "id") Application application,
             @Valid @RequestBody ApplicationRejectForm applicationRejectForm) {
         Student student = application.getStudent();
+        ApplicationStatus status = ApplicationStatus.valueOf(EntityUtil.getCode(application.getStatus()));
 
-        if (UserUtil.isStudentRepresentative(user, student)) {
+        if (UserUtil.isStudentRepresentative(user, student) && ApplicationStatus.AVALDUS_STAATUS_ESIT.equals(status)) {
             applicationService.reject(application, applicationRejectForm);
-        } else if (UserUtil.isSame(user, student) || UserUtil.isSchoolAdmin(user, student.getSchool())) {
+        } else if (UserUtil.isSchoolAdmin(user, student.getSchool()) && ApplicationStatus.AVALDUS_STAATUS_YLEVAAT.equals(status)) {
             applicationService.reject(application, applicationRejectForm);
         } else {
-            throw new RuntimeException("User has no rights to reject this application");
+            throw new HoisBusinessRuleException(String.format("user %s is not allowed to reject application %d with status %s", user.getUsername(), application.getId(), status));
         }
 
-        return get(application);
+        return get(user, application);
     }
+
+    private static void checkUpdateBusinessRules(HoisUserDetails user, Application application, ApplicationForm applicationForm) {
+        UserUtil.assertSameSchool(user, application.getStudent().getSchool());
+
+        Student student = application.getStudent();
+        School school = application.getStudent().getSchool();
+        ApplicationStatus status = ApplicationStatus.valueOf(EntityUtil.getCode(application.getStatus()));
+
+
+        switch (status) {
+        case AVALDUS_STAATUS_KOOST:
+            if ((UserUtil.isAdultStudent(user, student) || UserUtil.isSchoolAdmin(user, school)) && (applicationForm.getStatus().equals(ApplicationStatus.AVALDUS_STAATUS_KOOST.name())
+                    || applicationForm.getStatus().equals(ApplicationStatus.AVALDUS_STAATUS_ESIT.name()))) {
+                break;
+            } else if (!UserUtil.isAdultStudent(user, student) && (applicationForm.getStatus().equals(ApplicationStatus.AVALDUS_STAATUS_KOOST.name()))) {
+                break;
+            } else if (UserUtil.isStudentRepresentative(user, student) && (applicationForm.getStatus().equals(ApplicationStatus.AVALDUS_STAATUS_ESIT.name()) ||
+                    applicationForm.getStatus().equals(ApplicationStatus.AVALDUS_STAATUS_TAGASI.name()))) {
+                break;
+            }
+        case AVALDUS_STAATUS_ESIT:
+            //fallthrough
+        case AVALDUS_STAATUS_YLEVAAT:
+            if(UserUtil.isSchoolAdmin(user, school)
+                    && (applicationForm.getStatus().equals(ApplicationStatus.AVALDUS_STAATUS_YLEVAAT.name())) || applicationForm.getStatus().equals(ApplicationStatus.AVALDUS_STAATUS_TAGASI.name())) {
+                break;
+            }
+            //fallthrough
+        case AVALDUS_STAATUS_KINNITAM:
+            //fallthrough
+        case AVALDUS_STAATUS_KINNITATUD:
+            //fallthrough
+        case AVALDUS_STAATUS_TAGASI:
+            throw new HoisBusinessRuleException(String.format("user %s is not allowed to update application %d with status %s", user.getUsername(), application.getId(), status.name()));
+        default:
+            break;
+        }
+    }
+
 
 }
